requirements.txt
./
autograd==1.2
Flask==1.0.2
Flask-RESTful==0.3.6
future==0.16.0
gunicorn==19.8.1
lime==0.1.1.29
matplotlib==2.2.2
numpy==1.14.2
Pillow==4.2.1
requests==2.18.4
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==0.19.1
six==1.10.0
sklearn==0.0
tensorflow==1.6.0
text-unidecode==1.2
urllib3==1.22
